+++ Ork/Tests/Formatting/IgnoreCase.php
@@ -1,0 +1,50 @@
+<?php
+
+/**
+ * Sample class demonstrating sorting issues between ASCII and English.
+ *
+ * @package   Ork.Tests.Formatting
+ * @copyright sommeone
+ */
+
+namespace Ork\Tests\Formatting;
+
+/**
+ * Sample class demonstrating sorting issues between ASCII and English.
+ */
+class IgnoreCase
+{
+
+    /**
+     * Member var doc comment
+     *
+     * @var object[]
+     */
+    protected $boxes;
+
+    /**
+     * Member var doc comment
+     *
+     * @var object[]
+     */
+    protected $boxModels;
+
+    /**
+     * Function comment
+     *
+     * @return void
+     */
+    protected function getBoxes()
+    {
+    }
+
+    /**
+     * Function comment
+     *
+     * @return void
+     */
+    protected function getBoxModels()
+    {
+    }
+
+}